Convert profiler IDL namespace from RadarDDS to ProfilerDDS.
Change the type for the profiler DDS I/Q data from short to float, so that normalized coherently integrated values will retain their dynamic range.

git-svn-id: http://svn.eol.ucar.edu/svn/eol/common/trunk/ascope@7684 da89c675-a4d7-0310-8a15-9e7546503515

+++ AScope.cpp
@@ -633,7 +633,7 @@ AScope::newTSItemSlot(AScope::TimeSeries pItem) {
 
 		// extract the time series from the DDS sample
 		for (int t = 0; t < blockSize; t++) {
-            const short* ts = pItem.IQbeams[t];
+            const float* ts = pItem.IQbeams[t];
 			I[t] = ts[_gateChoice * 2];
 			Q[t] = ts[_gateChoice * 2 + 1];
 		}
